backend/requirements.txt
fastapi
uvicorn[standard]
httpx
websockets
python-multipart
aiofiles
markdown
pyyaml
psutil
